+++ client/tx/lcd/rest.go
@@ -1,0 +1,13 @@
+package lcd
+
+import (
+	"github.com/gorilla/mux"
+	"github.com/irisnet/irishub/client/context"
+	"github.com/irisnet/irishub/codec"
+)
+
+// RegisterRoutes - Central function to define routes that get registered by the main application
+func RegisterRoutes(cliCtx context.CLIContext, r *mux.Router, cdc *codec.Codec) {
+	r.HandleFunc("/tx/sign", SignTxRequestHandlerFn(cdc, cliCtx)).Methods("POST")
+	r.HandleFunc("/tx/broadcast", BroadcastTxRequestHandlerFn(cdc, cliCtx)).Methods("POST")
+}
